Added content
Header, footer, and title of break.
Next step is to add a visual for timer counting down, though I feel that might defeat the purpose of the app. The idea is a bar, or circle, that fills/empties as the timer counts down. This may or may not come into fruition.